Remove .idea files and update README and CMakeLists.txt files
This commit involves the removal of irrelevant .idea files from the repository. Moreover, the README.md file has been updated to change the instructions on how to run the HostExec.exe file. The CMakeLists.txt file in the Host directory is also updated to add the reference to 'helper.h'.