Add default root user to the database during server setup
Fix has been implemented to ensure uniqueness of usernames in the database. This was achieved by integrating the addUser() function within the server constructor to add a default user 'root' at the start, and throwing an error if it fails. This information is also communicated in the README for client usage instructions.