Add `--set-cwd` command line flag for server
Implemented the command line flag `--set-cwd` for the server application, which allows users to specify the current working directory for the server process. The server's usage message and the FLAGS_FOR_SERVER.md file were updated accordingly. A new method named `setCwd` was also added to handle the operation of setting the current directory.